+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+- Performance rule for Nette\Database\Table\Selection::fetch
+- Performance rule for checking calls in conditions
+- Performance rule for not use count($array) in conditions
+- Rule to Enforce arrow function when it contains only single return expression
+
+## [0.8.1] - 2026-01-27
+### Changed
+- Turn off Neon DI attribute Inject check by default
+
+## [0.8.0] - 2026-01-26
+### Added
+- Support for Neon DI attribute Inject
+- Set PHP 8.1 as minimum version
+
+## [0.7.6] - 2025-10-17
+### Added
+- Support for PHP 8.4
+
+## [0.7.5] - 2025-05-20
+### Fixed
+- Named parameters in unused schema properties
+
 ## [0.7.4] - 2025-03-03
 ### Fixed
 - Searching Schema definition
@@ -72,7 +95,11 @@
 - Rule for checking trait context 
 - Phpstorm meta dynamic return types
 
-[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.4...HEAD
+[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.8.1...HEAD
+[0.8.1]: https://github.com/efabrica-team/phpstan-rules/compare/0.8.0...0.8.1
+[0.8.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.6...0.8.0
+[0.7.6]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.5...0.7.6
+[0.7.5]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.4...0.7.5
 [0.7.4]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.3...0.7.4
 [0.7.3]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.2...0.7.3
 [0.7.2]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.1...0.7.2